[BEAM-3966] Move functional utilities into shared module
Some Java 8 functional utilities currently live under sdks/java/harness
but need to be shared more widely. This moves them into
sdks/java/fn-execution (but not into a new top-level module).
