+++ sdks/java/fn-execution/src/main/java/org/apache/beam/sdk/fn/function/CloseableThrowingConsumer.java
@@ -15,18 +15,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.apache.beam.sdk.fn.function;
 
-package org.apache.beam.fn.harness.fn;
-
-import java.util.function.Consumer;
-
-/**
- * A {@link Consumer} which can throw {@link Exception}s.
- *
- * <p>Used to expand the allowed set of method references to be used by Java 8
- * functional interfaces.
- */
-@FunctionalInterface
-public interface ThrowingConsumer<T> {
-  void accept(T t) throws Exception;
+/** A {@link ThrowingConsumer} that can be closed. */
+public interface CloseableThrowingConsumer<T> extends AutoCloseable, ThrowingConsumer<T> {
 }
