+++ sdks/java/fn-execution/src/main/java/org/apache/beam/sdk/fn/function/ThrowingBiConsumer.java
@@ -16,7 +16,18 @@
  * limitations under the License.
  */
 
+package org.apache.beam.sdk.fn.function;
+
+import java.util.function.BiConsumer;
+
 /**
- * Java 8 functional interface extensions.
+ * A {@link BiConsumer} which can throw {@link Exception}s.
+ *
+ * <p>Used to expand the allowed set of method references to be used by Java 8
+ * functional interfaces.
  */
-package org.apache.beam.fn.harness.fn;
+@FunctionalInterface
+public interface ThrowingBiConsumer<T1, T2> {
+  void accept(T1 t1, T2 t2) throws Exception;
+}
+
