+++ sdks/java/fn-execution/src/main/java/org/apache/beam/sdk/fn/function/ThrowingBiFunction.java
@@ -16,15 +16,17 @@
  * limitations under the License.
  */
 
-package org.apache.beam.fn.harness.fn;
+package org.apache.beam.sdk.fn.function;
+
+import java.util.function.BiFunction;
 
 /**
- * A {@link Runnable} which can throw {@link Exception}s.
+ * A {@link BiFunction} which can throw {@link Exception}s.
  *
  * <p>Used to expand the allowed set of method references to be used by Java 8
  * functional interfaces.
  */
 @FunctionalInterface
-public interface ThrowingRunnable {
-  void run() throws Exception;
+public interface ThrowingBiFunction<T1, T2, T3> {
+  T3 apply(T1 t1, T2 t2) throws Exception;
 }
