+++ sdks/java/fn-execution/src/main/java/org/apache/beam/sdk/fn/function/package-info.java
@@ -16,17 +16,5 @@
  * limitations under the License.
  */
 
-package org.apache.beam.fn.harness.fn;
-
-import java.util.function.Function;
-
-/**
- * A {@link Function} which can throw {@link Exception}s.
- *
- * <p>Used to expand the allowed set of method references to be used by Java 8
- * functional interfaces.
- */
-@FunctionalInterface
-public interface ThrowingFunction<T1, T2> {
-  T2 apply(T1 value) throws Exception;
-}
+/** Java 8 functional interface extensions. */
+package org.apache.beam.sdk.fn.function;
